# .migration-backup/artifacts/api-server/src/routes/payment.ts
import { Router } from "express";
import { randomUUID } from "crypto";
import { db, ordersTable, configServersTable, userPlansTable } from "@workspace/db";
import { eq, and } from "drizzle-orm";
import { InitiatePaymentBody } from "@workspace/api-zod";
import path from "path";
import { downloadConfigFile } from "../lib/storage";

const router = Router();

const PAYFLOW_BASE = "https://payflow.top/api/v2";
const PAYFLOW_API_KEY = process.env.PAYFLOW_API_KEY ?? "";
const PAYFLOW_API_SECRET = process.env.PAYFLOW_API_SECRET ?? "";
const PAYFLOW_ACCOUNT_ID = Number(process.env.PAYFLOW_ACCOUNT_ID ?? "0");

function payflowHeaders() {
  return {
    "X-API-Key": PAYFLOW_API_KEY,
    "X-API-Secret": PAYFLOW_API_SECRET,
    "Content-Type": "application/json",
  };
}

function normalizePhone(raw: string): string {
  const digits = raw.replace(/\D/g, "");
  if (digits.startsWith("0") && digits.length === 10) return `254${digits.slice(1)}`;
  if (digits.startsWith("254") && digits.length === 12) return digits;
  if (digits.startsWith("7") && digits.length === 9) return `254${digits}`;
  return digits;
}

function expiryFromDuration(duration: string): Date {
  const now = new Date();
  if (duration === "daily") now.setDate(now.getDate() + 1);
  else if (duration === "weekly") now.setDate(now.getDate() + 7);
  else now.setMonth(now.getMonth() + 1);
  return now;
}

async function autoFulfillOrder(orderId: string, logger: typeof console) {
  try {
    const [order] = await db.select().from(ordersTable).where(eq(ordersTable.id, orderId)).limit(1);
    if (!order || order.configUrl) return;

    const [server] = await db
      .select()
      .from(configServersTable)
      .where(
        and(
          eq(configServersTable.network, order.network),
          eq(configServersTable.appType, order.appType),
          eq(configServersTable.duration, order.duration),
          eq(configServersTable.status, "active")
        )
      )
      .limit(1);

    if (!server) {
      logger.warn?.(`No matching config server for order ${orderId}`);
      return;
    }

    // Verify file exists in Supabase Storage
    try {
      await downloadConfigFile(server.filename);
    } catch {
      logger.warn?.(`Config file missing in storage for server ${server.id}`);
      return;
    }

    const configUrl = `/api/orders/${orderId}/download`;
    const ext = path.extname(server.originalName).toLowerCase();

    await db.update(ordersTable)
      .set({ status: "completed", configUrl })
      .where(eq(ordersTable.id, orderId));

    const existing = await db.select().from(userPlansTable).where(eq(userPlansTable.orderId, orderId)).limit(1);
    if (existing.length === 0) {
      await db.insert(userPlansTable).values({
        id: randomUUID(),
        orderId,
        network: order.network,
        planName: server.serverName,
        planType: server.planType,
        duration: order.duration,
        appType: order.appType,
        deviceId: order.deviceId,
        phone: order.phone,
        expiryDate: expiryFromDuration(order.duration),
        status: "active",
        configUrl,
        fileExtension: ext,
      });
    }

    logger.info?.(`Auto-fulfilled order ${orderId} with config ${server.serverName}`);
  } catch (err) {
    logger.error?.(`Auto-fulfill failed for order ${orderId}: ${err}`);
  }
}

router.post("/initiate", async (req, res) => {
  const parsed = InitiatePaymentBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Invalid request body", details: parsed.error.issues });
    return;
  }

  const { phone, amount, orderId } = parsed.data;
  const reference = `NETCO-${randomUUID().slice(0, 8).toUpperCase()}`;
  const phoneFormatted = normalizePhone(phone);

  req.log.info({ phone: phoneFormatted, amount, orderId, reference }, "Initiating PayFlow STK push");

  try {
    const body = {
      payment_account_id: PAYFLOW_ACCOUNT_ID,
      phone: phoneFormatted,
      amount: Number(amount),
      reference,
      description: `NETCO VPN Config — Order ${orderId}`,
    };

    const pfRes = await fetch(`${PAYFLOW_BASE}/stkpush.php`, {
      method: "POST",
      headers: payflowHeaders(),
      body: JSON.stringify(body),
    });

    const pfData = (await pfRes.json()) as {
      success: boolean;
      message?: string;
      checkout_request_id?: string;
      data?: { checkout_request_id?: string };
    };

    req.log.info({ pfData, status: pfRes.status }, "PayFlow STK response");

    if (!pfRes.ok || !pfData.success) {
      res.status(502).json({
        error: "Payment gateway error",
        message: pfData.message ?? "STK Push failed. Please try again.",
      });
      return;
    }

    const checkoutRequestId =
      pfData.checkout_request_id ?? pfData.data?.checkout_request_id ?? `WS_CO_${Date.now()}`;

    await db
      .update(ordersTable)
      .set({ paymentReference: reference, status: "pending" })
      .where(eq(ordersTable.id, orderId));

    res.json({
      success: true,
      reference,
      checkoutRequestId,
      message: `M-Pesa STK Push sent to ${phone}. Enter your PIN on your phone.`,
    });
  } catch (err) {
    req.log.error({ err }, "PayFlow request failed");
    res.status(503).json({ error: "Payment service unreachable. Please try again shortly." });
  }
});

router.get("/status/:reference", async (req, res) => {
  const { reference } = req.params;

  try {
    const pfRes = await fetch(`${PAYFLOW_BASE}/status.php`, {
      method: "POST",
      headers: payflowHeaders(),
      body: JSON.stringify({ checkout_request_id: reference }),
    });

    const pfData = (await pfRes.json()) as {
      success: boolean;
      message?: string;
      status?: string;
      data?: {
        status?: string;
        transaction_code?: string;
        amount?: number;
        completed_at?: string;
      };
    };

    req.log.info({ pfData, reference }, "PayFlow status check");

    if (!pfRes.ok || !pfData.success) {
      res.json({ reference, status: "pending", message: pfData.message ?? null, completedAt: null });
      return;
    }

    const rawStatus = (pfData.data?.status ?? pfData.status ?? "pending").toLowerCase();

    let mappedStatus: "pending" | "completed" | "failed" | "cancelled";
    if (rawStatus === "completed" || rawStatus === "success") {
      mappedStatus = "completed";
    } else if (rawStatus === "failed" || rawStatus === "error") {
      mappedStatus = "failed";
    } else if (rawStatus === "cancelled") {
      mappedStatus = "cancelled";
    } else {
      mappedStatus = "pending";
    }

    let configUrl: string | null = null;

    if (mappedStatus === "completed") {
      const [order] = await db
        .select()
        .from(ordersTable)
        .where(eq(ordersTable.paymentReference, reference))
        .limit(1);

      if (order) {
        if (order.status !== "completed") {
          await autoFulfillOrder(order.id, req.log as typeof console);
        }
        const [freshOrder] = await db.select().from(ordersTable).where(eq(ordersTable.id, order.id)).limit(1);
        configUrl = freshOrder?.configUrl ?? null;
      }
    }

    res.json({
      reference,
      status: mappedStatus,
      message: pfData.message ?? null,
      completedAt: pfData.data?.completed_at ?? null,
      transactionCode: pfData.data?.transaction_code ?? null,
      configUrl,
    });
  } catch (err) {
    req.log.error({ err, reference }, "PayFlow status check failed");
    res.json({ reference, status: "pending", message: null, completedAt: null, configUrl: null });
  }
});

export default router;