# .migration-backup/lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETCO VPN Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputAppType } from './orderInputAppType';
import type { OrderInputDuration } from './orderInputDuration';

export interface OrderInput {
  packageId: string;
  network: string;
  duration: OrderInputDuration;
  appType: OrderInputAppType;
  deviceId: string;
  phone: string;
  amount?: number;
}