# .migration-backup/lib/api-zod/src/generated/types/packageCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETCO VPN Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PackagePlan } from './packagePlan';

export interface PackageCategory {
  id: string;
  name: string;
  plans: PackagePlan[];
}