# .migration-backup/lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETCO VPN Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInput {
  phone: string;
  amount: number;
  orderId: string;
}