# .migration-backup/lib/api-zod/src/generated/types/revenueByMonth.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETCO VPN Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RevenueByMonth {
  month: string;
  revenue: number;
  orders: number;
}